move update logic into hook
