+++ src/graphql-queries/index.js
@@ -1,4 +1,5 @@
 export { MY_WORDLIST } from './myWordlist';
+export { MY_WORDLIST_CATEGORIES } from './myWordlistCategories';
 export { MY_WORDLIST_CREATE } from './myWordlistCreate';
 export { WORDLIST_ENTRIES_CREATE } from './wordlistEntriesCreate';
 export { WORDLIST_ENTRY_DELETE } from './wordlistEntryDelete';
